Revert "change search design"
