--- table_customisation/src/App.css
@@ -1,9 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-
-body {
-  margin: 0;
-  padding: 0;
-  background-color: #f3f4f6;
-}
