--- table_customisation/src/App.js
@@ -1,12 +1,0 @@
-import React from "react";
-import CarTable from "./CarTable";
-import "./App.css";
-const App = () => {
-  return (
-    <div>
-      <CarTable />
-    </div>
-  );
-};
-
-export default App;
